test.c is written.

+++ encrypt/Encrypt.h
@@ -19,19 +19,19 @@
 // void keygen(int key);
 
 // This is expand function which expands 16 bit to 24 bit.
-int expand(short int half);
+int expand(int half);
 
 // This is for substitution of 24 bit to 16 bit.
-short int subst(int sub);
+int subst(int sub);
 
 // This is for permutation after substitution. rearrange bits to get a short int
-short int perm(short int sub);
+int perm(int sub);
 
 // This gives one hot at the given index.
 int onehot(int index);
 
 // This prints the int in form of bits.
-void printbits(int num);
+void printbits(unsigned int num);
 
 
 int fiestel(int input);
